Refactor packages and verify user from email in user service
